Update GNU Emacs to 31.1

+++ bin/install/emacs-install.sh
@@ -2,15 +2,28 @@
 set -e
 set -x
 
-d="/usr/local/stow/emacs-30.2-gtk1"
+d="/usr/local/stow/emacs-31.1-deb13+lucid1"
+
+if [[ -e "$d" ]] ; then
+    echo "Cowardly refusing to overwrite directory $d" >&2
+    exit 1
+fi
 
 make install "prefix=$d"
 cd "$d/bin"
-mv -v ctags ctags-emacs
-mv -v etags etags-emacs
+if [ -e ctags ] ; then
+    mv -v ctags ctags-emacs
+fi
+if [ -e etags ] ; then
+    mv -v etags etags-emacs
+fi
 cd ../share/man/man1
-mv -v ctags.1.gz ctags-emacs.1.gz
-mv -v etags.1.gz etags-emacs.1.gz
-echo '.so man1/etags-emacs.1' | gzip -c | tee ctags-emacs.1.gz > /dev/null
+if [ -e ctags.1.gz ] ; then
+    mv -v ctags.1.gz ctags-emacs.1.gz
+    echo '.so man1/etags-emacs.1' | gzip -c | tee ctags-emacs.1.gz > /dev/null
+fi
+if [ -e etags.1.gz ] ; then
+    mv -v etags.1.gz etags-emacs.1.gz
+fi
 cd
 sudo chown -R root:root "$d"
